Initial Code Commit for the Project Structure
We will want to rip out the majority of this work.  We just want a good starting point for now.  For the sake of time, we have committed the sample project our professor has kicked us off with as the structure for our application will be very similar.  We expect this initial commit to change drastically with future changesets.